pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "codeassistant"
version = "0.1.0"
description = "Local code assistant CLI with per-project context and cost tracking"
authors = [{name = "Anupam Tripathi"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
codeassistant = "codeassistant.cli:main"